+++ application/account/service/account_service.go
@@ -6,7 +6,10 @@ import (
 	"github.com/kiosanim/pismo-code-assessment/application/account/dto"
 	"github.com/kiosanim/pismo-code-assessment/application/account/mapper"
 	"github.com/kiosanim/pismo-code-assessment/internal/core/cache"
+	"github.com/kiosanim/pismo-code-assessment/internal/core/contextutils"
 	coreerr "github.com/kiosanim/pismo-code-assessment/internal/core/errors"
+	"github.com/kiosanim/pismo-code-assessment/internal/core/factory"
+	"github.com/kiosanim/pismo-code-assessment/internal/core/lock"
 	"github.com/kiosanim/pismo-code-assessment/internal/core/logger"
 	"github.com/kiosanim/pismo-code-assessment/internal/domains/account"
 )
@@ -15,46 +18,74 @@ type AccountService struct {
 	accountRepository account.AccountRepository
 	cache             cache.CacheRepository
 	componentName     string
+	locker            lock.DistributedLockManager
 	log               logger.Logger
 }
 
-func NewAccountService(repository account.AccountRepository, cache cache.CacheRepository, log logger.Logger) *AccountService {
-	accountService := AccountService{accountRepository: repository, cache: cache, log: log}
-	accountService.componentName = logger.ComponentNameFromStruct(accountService)
-	return &accountService
+func NewAccountService(factory factory.Factory) *AccountService {
+	return &AccountService{
+		componentName:     "AccountService",
+		accountRepository: factory.AccountRepository(),
+		cache:             factory.CacheRepository(),
+		locker:            factory.DistributedLockManager(),
+		log:               factory.Log(),
+	}
 }
 
 func (a *AccountService) FindByID(ctx context.Context, request dto.FindAccountByIdRequest) (*dto.FindAccountByIdResponse, error) {
-	a.log.Debug(a.componentName+".FindByID", "request", request)
+	traceID := contextutils.GetTraceID(ctx)
+	a.log.Debug(a.componentName+".FindByID", "request", request, "x_trace_id", traceID)
 	if request.AccountID <= 0 {
-		return nil, coreerr.InvalidParametersError
+		err := coreerr.InvalidParametersError
+		a.log.Warn(a.componentName+".FindByID", "error", err, "x_trace_id", traceID)
+		return nil, err
 	}
 	output, err := a.accountRepository.FindByID(ctx, request.AccountID)
 	if err != nil {
+		a.log.Warn(a.componentName+".FindByID", "error", err, "x_trace_id", traceID)
 		return nil, err
 	}
 	if output == nil {
-		return nil, coreerr.AccountNotFoundError
+		err := coreerr.AccountNotFoundError
+		a.log.Warn(a.componentName+".FindByID", "error", err, "x_trace_id", traceID)
+		return nil, err
 	}
 	response := mapper.FindEntityToResponse(output)
 	return response, nil
 }
 
 func (a *AccountService) Create(ctx context.Context, request dto.CreateAccountRequest) (*dto.CreateAccountResponse, error) {
-	a.log.Debug(a.componentName+".Create", "request", request)
+	traceID := contextutils.GetTraceID(ctx)
+	a.log.Debug(a.componentName+".Create", "request", request, "x_trace_id", traceID)
 	if request.DocumentNumber == "" || account.IsValidDocumentNumber(request.DocumentNumber) != nil {
-		return nil, coreerr.InvalidParametersError
+		err := coreerr.InvalidParametersError
+		a.log.Warn(a.componentName+".Create", "error", err, "x_trace_id", traceID)
+		return nil, err
 	}
 	accountByDocumentNumber, err := a.accountRepository.FindByDocumentNumber(ctx, request.DocumentNumber)
 	if err != nil && !errors.Is(err, coreerr.AccountNotFoundError) {
+		a.log.Warn(a.componentName+".Create", "error", err, "x_trace_id", traceID)
 		return nil, err
 	}
 	if accountByDocumentNumber != nil {
-		return nil, coreerr.AccountAlreadyExistsForDocumentNumberError
+		err := coreerr.AccountAlreadyExistsForDocumentNumberError
+		a.log.Warn(a.componentName+".Create", "error", err, "x_trace_id", traceID)
+		return nil, err
 	}
 	accountRequest := mapper.CreateDTOToEntity(request)
+	lck, err := a.locker.WaitToLockUsingDefaultTimeConfiguration(ctx, lock.AccountCreationLockKey)
+	if err != nil {
+		a.log.Warn(a.componentName+".Create", "error", err, "x_trace_id", traceID)
+		return nil, err
+	}
 	output, err := a.accountRepository.Save(ctx, accountRequest)
 	if err != nil {
+		a.log.Warn(a.componentName+".Create", "error", err, "x_trace_id", traceID)
+		return nil, err
+	}
+	err = a.locker.Unlock(ctx, lck)
+	if err != nil {
+		a.log.Warn(a.componentName+".Create", "error", err, "x_trace_id", traceID)
 		return nil, err
 	}
 	response := mapper.CreateEntityToResponse(output)