+++ interfaces/http/router/router_factory.go
@@ -2,12 +2,10 @@ package router
 
 import (
 	"github.com/gin-gonic/gin"
+	acc "github.com/kiosanim/pismo-code-assessment/application/account/service"
+	tra "github.com/kiosanim/pismo-code-assessment/application/transaction/service"
 	"github.com/kiosanim/pismo-code-assessment/internal/core/logger"
 	"github.com/kiosanim/pismo-code-assessment/internal/infra/factory"
-)
-
-const (
-	ComponentName = "RouterFactory"
 )
 
 func NewRouterFactory(appFactory factory.AppFactory, log logger.Logger) *gin.Engine {
@@ -19,19 +17,19 @@ func NewRouterFactory(appFactory factory.AppFactory, log logger.Logger) *gin.Eng
 	if transactionRepo == nil {
 		panic("Transaction Repository not initialized")
 	}
-	accountSvc := appFactory.AccountService()
+	accountSvc := acc.NewAccountService(&appFactory)
 	if accountSvc == nil {
 		panic("Account Service not initialized")
 	}
-	transactionSvc := appFactory.TransactionService()
+	transactionSvc := tra.NewTransactionService(&appFactory)
 	if transactionSvc == nil {
 		panic("Transaction Service not initialized")
 	}
-	accountHandler := appFactory.AccountHandler()
+	accountHandler := appFactory.AccountHandler(accountSvc)
 	if accountHandler == nil {
 		panic("Account Handler not initialized")
 	}
-	transactionHandler := appFactory.TransactionHandler()
+	transactionHandler := appFactory.TransactionHandler(transactionSvc)
 	if transactionHandler == nil {
 		panic("Transaction Handler not initialized")
 	}