+++ internal/infra/database/model/operation_type_model.go
@@ -1,9 +1,6 @@
 package model
 
-import "github.com/uptrace/bun"
-
 type OperationTypeModel struct {
-	bun.BaseModel   `bun:"table:operation_types,alias:t"`
 	OperationTypeID int64  `bun:"operation_type_id,pk,autoincrement"` // Unique identifier of an OperationType
 	Description     string `bun:"description,notnull"`
 }