+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] 2026-02-05
+- Add distributed lock to transactions
+- Add warning and debug messages
+- Add x-trace-id in log messages
+
 ## [0.6.0] 2026-02-04
 - Add sources already commited
 - Create a redis repository